minor changes bugfixing for live env. issue with dividing a float.

+++ queries.py
@@ -136,12 +136,12 @@ class CalculateViewData():
         viewsDelta = float(recordA.views - recordB.views)
         
         if timeDeltaSeconds != 0.0:
-        # calculate the average speed since the last check in vps
-            recordASpeed = float(viewsDelta / timeDeltaSeconds) * 60 * 60
-        
-        # calculate the change in speed
+            # calculate the average speed since the last check in vps
+            recordASpeed = float((viewsDelta / timeDeltaSeconds) * 60 * 60)
+                    
+            # calculate the change in speed
             speedDelta = recordASpeed - recordBSpeed
-            viewsAcceleration = speedDelta / timeDeltaSeconds * 60 * 60
+            viewsAcceleration = float((speedDelta / timeDeltaSeconds) * 60 * 60)
         
         return viewsSpeed, viewsAcceleration
